Adjust the alignment and sort key in output results. (#220)
* Update poller_mactrack.php

Formating the source code

* Update mactrack_functions.php

Formatting the code

* Update mactrack_functions.php

Formatting the code

* Update mactrack_view_macs.php

1. Fix  the IP address's delimiter.
2. Formatting the code.

* Update mactrack_macauth.php

Formatting the code

* Update mactrack_macauth.php

Formatting the code

* Update mactrack_view_arp.php

Adjust the text filed with the left alignment for easy readable.

* Update mactrack_macauth.php

The default to view the authorized mac address with descending sort is more useful.

* Update mactrack_view_macs.php

1. Adjust the text filed with the left alignment for easy readable.

* Update mactrack_cisco.php

The Cisco default vlan are 1002-1005 (4 vlans).

* Update mactrack_cisco.php

In function get_IOS_dot1dTpFdbEntry_ports, it doesn't need to run collect ifIndexes information twice.